+++ src/main/java/org/akhq/controllers/SchemaController.java
@@ -4,6 +4,7 @@ import io.confluent.kafka.schemaregistry.client.rest.exceptions.RestClientExcept
 import io.micronaut.context.annotation.Value;
 import io.micronaut.http.HttpRequest;
 import io.micronaut.http.HttpResponse;
+import io.micronaut.http.HttpStatus;
 import io.micronaut.http.annotation.*;
 import io.micronaut.security.annotation.Secured;
 import io.swagger.v3.oas.annotations.Operation;
@@ -11,6 +12,7 @@ import org.akhq.configs.Role;
 import org.akhq.middlewares.SchemaComparator;
 import org.akhq.models.Schema;
 import org.akhq.models.TopicSchema;
+import org.akhq.modules.KafkaModule;
 import org.akhq.repositories.SchemaRegistryRepository;
 import org.akhq.utils.Pagination;
 import org.akhq.utils.ResultPagedList;
@@ -35,6 +37,8 @@ public class SchemaController extends AbstractController {
     public SchemaController(SchemaRegistryRepository schemaRepository) {
         this.schemaRepository = schemaRepository;
     }
+    @Inject
+    private KafkaModule kafkaModule;
 
     @Get("api/{cluster}/schema")
     @Operation(tags = {"schema registry"}, summary = "List all schemas")
@@ -44,11 +48,16 @@ public class SchemaController extends AbstractController {
         Optional<String> search,
         Optional<Integer> page
     ) throws IOException, RestClientException, ExecutionException, InterruptedException {
+        if(kafkaModule.clusterExists(cluster)) {
 
-        URIBuilder uri = URIBuilder.fromURI(request.getUri());
-        Pagination pagination = new Pagination(pageSize, uri, page.orElse(1));
+            URIBuilder uri = URIBuilder.fromURI(request.getUri());
+            Pagination pagination = new Pagination(pageSize, uri, page.orElse(1));
 
-        return ResultPagedList.of(this.schemaRepository.list(cluster, pagination, search));
+            return ResultPagedList.of(this.schemaRepository.list(cluster, pagination, search));
+        } else {
+            HttpResponse.status(HttpStatus.NOT_FOUND);
+            return null;
+        }
     }
 
     @Get("api/{cluster}/schema/topic/{topic}")
@@ -58,16 +67,21 @@ public class SchemaController extends AbstractController {
         String cluster,
         String topic
     ) throws IOException, RestClientException {
-        List<Schema> schemas = this.schemaRepository.listAll(cluster, Optional.empty());
+        if(!kafkaModule.clusterExists(cluster)) {
+            List<Schema> schemas = this.schemaRepository.listAll(cluster, Optional.empty());
 
-        return new TopicSchema(
-            schemas.stream()
-                .sorted(new SchemaComparator(topic, true))
-                .collect(Collectors.toList()),
-            schemas.stream()
-                .sorted(new SchemaComparator(topic, false))
-                .collect(Collectors.toList())
-        );
+            return new TopicSchema(
+                schemas.stream()
+                    .sorted(new SchemaComparator(topic, true))
+                    .collect(Collectors.toList()),
+                schemas.stream()
+                    .sorted(new SchemaComparator(topic, false))
+                    .collect(Collectors.toList())
+            );
+        } else {
+            HttpResponse.status(HttpStatus.NOT_FOUND);
+            return null;
+        }
     }
 
     @Secured(Role.ROLE_REGISTRY_INSERT)
@@ -77,32 +91,47 @@ public class SchemaController extends AbstractController {
         String cluster,
         @Body Schema schema
     ) throws Throwable {
-        if (this.schemaRepository.exist(cluster, schema.getSubject())) {
-            throw new IllegalArgumentException("Subject '" + schema.getSubject() + "' already exits");
-        }
+        if(kafkaModule.clusterExists(cluster)) {
+            if (this.schemaRepository.exist(cluster, schema.getSubject())) {
+                throw new IllegalArgumentException("Subject '" + schema.getSubject() + "' already exits");
+            }
 
-        return registerSchema(cluster, schema);
+            return registerSchema(cluster, schema);
+        } else {
+            HttpResponse.status(HttpStatus.NOT_FOUND);
+            return null;
+        }
     }
 
     @Get("api/{cluster}/schema/{subject}")
     @Operation(tags = {"schema registry"}, summary = "Retrieve a schema")
     public Schema home(HttpRequest<?> request, String cluster, String subject) throws IOException, RestClientException {
-        return this.schemaRepository.getLatestVersion(cluster, subject);
+        if(kafkaModule.clusterExists(cluster)) {
+            return this.schemaRepository.getLatestVersion(cluster, subject);
+        } else {
+            HttpResponse.status(HttpStatus.NOT_FOUND);
+            return null;
+        }
     }
 
     @Secured(Role.ROLE_REGISTRY_UPDATE)
     @Post(value = "api/{cluster}/schema/{subject}")
     @Operation(tags = {"schema registry"}, summary = "Update a schema")
     public Schema updateSchema(String cluster, String subject, @Body Schema schema) throws Throwable {
-        if (!this.schemaRepository.exist(cluster, subject)) {
-            throw new IllegalArgumentException("Subject '" + subject + "' doesn't exits");
-        }
+        if(kafkaModule.clusterExists(cluster)) {
+            if (!this.schemaRepository.exist(cluster, subject)) {
+                throw new IllegalArgumentException("Subject '" + subject + "' doesn't exits");
+            }
 
-        if (!subject.equals(schema.getSubject())) {
-            throw new IllegalArgumentException("Invalid subject name '" + subject + "', doesn't egals '" + schema.getSubject() + "'");
-        }
+            if (!subject.equals(schema.getSubject())) {
+                throw new IllegalArgumentException("Invalid subject name '" + subject + "', doesn't egals '" + schema.getSubject() + "'");
+            }
 
-        return registerSchema(cluster, schema);
+            return registerSchema(cluster, schema);
+        } else {
+            HttpResponse.status(HttpStatus.NOT_FOUND);
+            return null;
+        }
     }
 
     private Schema registerSchema(String cluster, @Body Schema schema) throws IOException, RestClientException {
@@ -128,28 +157,42 @@ public class SchemaController extends AbstractController {
         String cluster,
         Integer id
     ) throws IOException, RestClientException, ExecutionException, InterruptedException {
-        return this.schemaRepository
-            .getById(cluster, id)
-            .orElse(null);
+        if(kafkaModule.clusterExists(cluster)) {
+            return this.schemaRepository
+                .getById(cluster, id)
+                .orElse(null);
+        } else {
+            HttpResponse.status(HttpStatus.NOT_FOUND);
+            return null;
+        }
     }
 
     @Get("api/{cluster}/schema/{subject}/version")
     @Operation(tags = {"schema registry"}, summary = "List all version for a schema")
     public List<Schema> versions(HttpRequest<?> request, String cluster, String subject) throws IOException, RestClientException {
-        return this.schemaRepository.getAllVersions(cluster, subject);
+        if(kafkaModule.clusterExists(cluster)) {
+            return this.schemaRepository.getAllVersions(cluster, subject);
+        } else {
+            HttpResponse.status(HttpStatus.NOT_FOUND);
+            return null;
+        }
     }
 
     @Secured(Role.ROLE_REGISTRY_DELETE)
     @Delete("api/{cluster}/schema/{subject}")
     @Operation(tags = {"schema registry"}, summary = "Delete a schema")
     public HttpResponse<?> delete(String cluster, String subject) throws IOException, RestClientException {
-        if (!this.schemaRepository.exist(cluster, subject)) {
-            throw new IllegalArgumentException("Subject '" + subject + "' doesn't exits");
+        if(kafkaModule.clusterExists(cluster)) {
+            if (!this.schemaRepository.exist(cluster, subject)) {
+                throw new IllegalArgumentException("Subject '" + subject + "' doesn't exits");
+            }
+
+            this.schemaRepository.delete(cluster, subject);
+
+            return HttpResponse.noContent();
+        } else {
+            return HttpResponse.status(HttpStatus.NOT_FOUND);
         }
-
-        this.schemaRepository.delete(cluster, subject);
-
-        return HttpResponse.noContent();
     }
 
     @Secured(Role.ROLE_REGISTRY_VERSION_DELETE)
@@ -160,8 +203,12 @@ public class SchemaController extends AbstractController {
         String subject,
         Integer version
     ) throws IOException, RestClientException {
-        this.schemaRepository.deleteVersion(cluster, subject, version);
+        if(kafkaModule.clusterExists(cluster)) {
+            this.schemaRepository.deleteVersion(cluster, subject, version);
 
-        return HttpResponse.noContent();
+            return HttpResponse.noContent();
+        } else {
+            return HttpResponse.status(HttpStatus.NOT_FOUND);
+        }
     }
 }
